+++ src/main/java/com/optimal/standard/persistence/model/ConstructionSystemMaterial.java
@@ -1,14 +1,6 @@
 package com.optimal.standard.persistence.model;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -20,25 +12,25 @@ import lombok.NoArgsConstructor;
 @Table(name = "construction_system_material")
 public class ConstructionSystemMaterial {
 
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
 
-  @ManyToOne
-  @JoinColumn(name = "material_id")
-  private Material material;
+    @ManyToOne
+    @JoinColumn(name = "material_id")
+    private Material material;
 
-  @ManyToOne
-  @JoinColumn(name = "construction_system_id")
-  private ConstructionSystem constructionSystem;
+    @ManyToOne
+    @JoinColumn(name = "construction_system_id")
+    private ConstructionSystem constructionSystem;
 
-  @Enumerated(EnumType.STRING)
-  private TypeOfUse typeOfUse;
+    @Enumerated(EnumType.STRING)
+    private TypeOfUse typeOfUse;
 
-  private Integer coefficient;
+    private Double coefficient;
 
-  private String coefficientDescription;
+    private String coefficientDescription;
 
-  private String materialDescription;
+    private String materialDescription;
 
 }
